chore(restructure): move saasshorts -> openshorts/saas/pipeline.py
Phase 1 step 8: relocate the SaaS UGC pipeline as a single module. The
plan calls for an internal split into research / scripting / media /
compositing / pipeline; doing that as one move keeps the change small,
ships the file into the right folder, and defers the function-level
split to a follow-up commit. Shim at saasshorts.py preserves existing
`from saasshorts import ...` calls. No direct test coverage for this
module; the openapi.json contract still passes. Tests stay 62/62 green.
